Move to for_each states with repo names as keys
Previously the states of the resources had the index of the repos in the
variable in them. This leads to a lot of force recreate when you remove
something from the middle since all the indexes afterwards get changed.

Now we use the repository name as the key for the repository states
which is also unique and more permanent.

Signed-off-by: Colin Wilk <colin.wilk@tum.de>